Run 2to3 on pystache code
It turns out we didn't do enough in f4255823; there are still other
incompatibilities in the pystache code. pystache actually handles these
issues with the 2to3 tool, which is run during setup.py. So the
root of the problem is not that pystache doesn't properly support Python
3, but rather that we aren't running 2to3 because we aren't installing
pystache the normal way (this was triggered by 688446c1, where we
changed the way pystache is downloaded and packaged into the .ankiaddon
file).

Fix this by running `2to3` on the pystache code before zipping it into
package.ankiaddon.