Allow re-adding seen words
I've fallen behind on study and now I have a lot of words in my
collection that I don't really know. I want to be able to re-add some of
these words to a new deck for a new prestudy project, creating entirely
new cards that don't have any review history.

Add Seen words to the word list in the UI. By default, they won't be
added, but you can click the box to add them.

There is a known bug in this implementation: if the existing note was
also generated by Chinese-Prestudy (or the chineseflashcards library),
the new note will have a matching GUID, and adding it will silently do
nothing (no new note will be added and the old note will not be updated
/ moved into the selected deck).